requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==25.12.0
celery==5.6.2
celery-types==0.24.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-parler==2.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
html-sanitizer==2.6.0
inflection==0.5.1
isort==7.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
lxml_html_clean==0.4.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytokens==0.3.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Unidecode==1.4.0
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
